[rename] : security 관련 클래스 파일 경로 수정
