+++ src/main/java/com/dnd/gongmuin/security/dto/Oauth2Response.java
@@ -1,4 +1,4 @@
-package com.dnd.gongmuin.auth.dto;
+package com.dnd.gongmuin.security.dto;
 
 public interface Oauth2Response {
 	String getProvider();
