+++ src/main/java/com/dnd/gongmuin/security/handler/CustomOauth2AuthenticationSuccessHandler.java
@@ -1,4 +1,4 @@
-package com.dnd.gongmuin.auth.handler;
+package com.dnd.gongmuin.security.handler;
 
 import java.io.IOException;
 
@@ -6,12 +6,12 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 
-import com.dnd.gongmuin.auth.dto.CustomOauth2User;
 import com.dnd.gongmuin.auth.service.AuthService;
 import com.dnd.gongmuin.common.exception.runtime.NotFoundException;
 import com.dnd.gongmuin.member.domain.Member;
 import com.dnd.gongmuin.member.exception.MemberErrorCode;
 import com.dnd.gongmuin.member.repository.MemberRepository;
+import com.dnd.gongmuin.security.dto.CustomOauth2User;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -30,12 +30,14 @@ public class CustomOauth2AuthenticationSuccessHandler implements AuthenticationS
 		Authentication authentication) throws IOException, ServletException {
 
 		CustomOauth2User customOauth2User = (CustomOauth2User)authentication.getPrincipal();
+		System.out.println("customOauth2User.getName() = " + customOauth2User.getName());
 
 		String socialEmail = customOauth2User.getEmail();
 		Member findmember = memberRepository.findBySocialEmail(socialEmail)
 			.orElseThrow(() -> new NotFoundException(MemberErrorCode.NOT_FOUND_MEMBER));
 
 		if (!isAuthStatusOld(findmember)) {
+			// TODO : 임시 JWT 토큰 발급 구현
 			response.sendRedirect("/additional-info");
 		} else {
 			// TODO : 기존회원 JWT 발급 및 페이지 리다이렉션 구현
